requirements.txt
# Core backend
fastapi==0.115.5
uvicorn[standard]==0.30.6
redis==5.0.7
pydantic==2.9.2
python-dotenv==1.0.1

# ✅ Stable Groq SDK + matching httpx
groq==0.5.0
httpx==0.27.0